feat: These are modifications to DSpace 5.5 required to implment authentication for the xmlui and rest and updates to the REST API.
This commit does not contain any Willamette-specific UI updates to the xmlui.  Authentication methods are CASAuthentication, RestAuthentication and WUSpecialGroups.  The xmlui is updated for CASAuthentication.  Quite a few updates to the REST API for DSpace 5.5 are included.  This update does not use the dspace/modules for war overlays or for additions, simply because a) this process seems unreliable or fragile, b) some updates to these directories require updates to pom.xml files with each release so we can't simply copy into these directories and overlay without review, 3) it tends to make it moderately more difficult to develop code even when it works.