bug: Unable to access DSpace objects as anonymous user.
The problem was caused by a null pointer exception that originiated in utility methods. As a result, the finally clause was aborting context and sending a 500 WebApplicationException.  The API client currently reponds to the 500 error by attempting to login.  Authorized users were therefore able to see public objects that were still hidden to anonymous users.

Issues: DSPACE-130