bug: Possible fix for null eperson creation during DSpace authentication.
It's not clear to me how CASAuthentication was being triggered with a null netid, but it looks like was happening.  As a result, SpeakerToLdap.getLdapAttribute was returning an empty map. CASAuthentication module was using this map to set eperson attributes to null values and adding the new eperson to the database.  Added a check for null or empty emal attribute that will prevent an empty eperson from being created.

Issues: DSPACE-133